:recycle: Move useLiveLocation into LocationContext
