+++ src/components/context/LocationContext.tsx
@@ -1,17 +1,54 @@
 import { Location } from '../../types';
 import React, { createContext, PropsWithChildren, useContext, useEffect, useState } from 'react';
-import { useInterval, useLiveLocation } from '../../hooks';
+import { useInterval } from '../../hooks';
 import useLocationsV1 from '../../services/locations_v1';
 import { calculateDistance } from '../../utils/location';
 import { Alert } from 'react-native';
 import { Magnetometer, MagnetometerMeasurement } from 'expo-sensors';
+import { Accuracy, LocationObject, LocationSubscription, watchPositionAsync } from 'expo-location';
 
-const LOCATION_UPLOAD_DELAY = 1000 * 10; // 10 seconds
-const LOCATION_UPLOAD_DISTANCE_INTERVAL = 10; // measured in metres
+const UPLOAD_DELAY = 1000 * 10; // 10 seconds
+const UPLOAD_DISTANCE_INTERVAL = 10; // measured in metres
 // TODO: implement this
-const LOCATION_UPLOAD_MAX_DELAY = 1000 * 60 * 5; // 5 minutes
+const UPLOAD_MAX_DELAY = 1000 * 60 * 5; // 5 minutes
 
 const MAGNETOMETER_UPDATE_INTERVAL = 50; // milliseconds
+
+const LOCATION_UPDATE_INTERVAL = 1000; // 1 second
+const LOCATION_DISTANCE_INTERVAL = 1; // measured in metres
+
+const useLiveLocation = () => {
+  const [location, setLocation] = useState<Location>();
+
+  const updateLocation = (data: LocationObject) => {
+    setLocation({
+      longitude: data.coords.longitude,
+      latitude: data.coords.latitude,
+      altitude: data.coords.altitude,
+    });
+  };
+
+  useEffect(() => {
+    let subscription: LocationSubscription;
+
+    (async () => {
+      subscription = await watchPositionAsync(
+        {
+          accuracy: Accuracy.High, // <= 10m accuracy
+          timeInterval: LOCATION_UPDATE_INTERVAL,
+          distanceInterval: LOCATION_DISTANCE_INTERVAL,
+        },
+        updateLocation
+      );
+    })();
+
+    return () => {
+      subscription.remove();
+    };
+  }, []);
+
+  return location;
+};
 
 const useMagneticHeading = () => {
   const [heading, setHeading] = useState<number>(0);
@@ -60,10 +97,7 @@ const LocationProvider: React.FC<PropsWithChildren> = ({ children }) => {
     const temp = location;
     const payload = { ...temp, timestamp: Date.now() };
 
-    if (
-      !lastLocation ||
-      calculateDistance(lastLocation, location) > LOCATION_UPLOAD_DISTANCE_INTERVAL
-    ) {
+    if (!lastLocation || calculateDistance(lastLocation, location) > UPLOAD_DISTANCE_INTERVAL) {
       locationsService.setLocation(payload).then((reason) => {
         if (reason) {
           Alert.alert(reason);
@@ -72,7 +106,7 @@ const LocationProvider: React.FC<PropsWithChildren> = ({ children }) => {
         }
       });
     }
-  }, LOCATION_UPLOAD_DELAY);
+  }, UPLOAD_DELAY);
 
   return (
     <LocationContext.Provider value={{ location, heading }}>{children}</LocationContext.Provider>
