+++ src/hooks/index.ts
@@ -1,3 +1,2 @@
-export { default as useLiveLocation } from './useLiveLocation';
 export { default as useColors } from './useColors';
 export { default as useInterval } from './useInterval';
